Ticket #1740
Message: Fixing the VS 2010 build. As part of this I converted the Contrib Solution and its projects to VS 2010. To get the WebService_vc10.vcxproj to build it was necessary to create a gsoap-[ds].lib from VS 2010. This new lib was added to sdlExtern as gsoap-[ds]_vc_10.lib to the respective folders lib_win32 and lib_x64. The WebService_vc10.vcxproj was updated to use this library instead of the gsoap-d_vc10.lib[ds].lib.

I also fixed the breakage in the Video Screener standalone application about not being able to find Version.hpp

Fixed the GPSIfc_vc10.vcxproj so it was not looking for the FvwClient that no longer exists.

Owners Code Review: No code was changed, just new solutions and projects created, but I did talk to Ben Marchant and Smanders about the new solution and projects and the new library name and they concurred.

git-svn-id: https://isrsvn.usurf.usu.edu/svn/isr/roots/Contrib/gsoap@17024 a06e8daa-702c-0410-a579-eb5e6c4fa09d

Signed-off-by: Scott M Anderson <smanders@sdl.usu.edu>